+++ pauli_tracker/src/pauli/tuple.rs
@@ -19,9 +19,9 @@ use super::Pauli;
 #[derive(Debug, Clone, Copy, Default, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct PauliTuple(
-    /// X part
-    pub bool,
     /// Z part
+    pub bool,
+    /// X part
     pub bool,
 );
 
@@ -38,12 +38,12 @@ impl Display for PauliTuple {
 
 impl Pauli for PauliTuple {
     const I: Self = Self(false, false);
-    const X: Self = Self(true, false);
+    const Z: Self = Self(true, false);
     const Y: Self = Self(true, true);
-    const Z: Self = Self(false, true);
+    const X: Self = Self(false, true);
 
-    fn new_product(x: bool, z: bool) -> Self {
-        Self(x, z)
+    fn new_product(z: bool, x: bool) -> Self {
+        Self(z, x)
     }
 
     fn add(&mut self, other: Self) {
@@ -52,58 +52,58 @@ impl Pauli for PauliTuple {
     }
 
     fn s(&mut self) {
-        self.1 ^= self.0;
+        self.0 ^= self.1;
     }
     fn h(&mut self) {
-        mem::swap(&mut self.0, &mut self.1);
+        mem::swap(&mut self.1, &mut self.0);
     }
     fn sh(&mut self) {
         // cf. stack impl
-        mem::swap(&mut self.0, &mut self.1);
-        self.1 ^= self.0;
+        self.h();
+        self.s();
     }
     fn hs(&mut self) {
         // cf. stack impl
-        self.1 ^= self.0;
-        mem::swap(&mut self.0, &mut self.1);
+        self.s();
+        self.h();
     }
     fn shs(&mut self) {
-        self.0 ^= self.1;
+        self.1 ^= self.0;
     }
 
     fn xpx(&mut self, other: &Self) {
-        self.0 ^= other.0;
-    }
-
-    fn xpz(&mut self, other: &Self) {
-        self.0 ^= other.1;
-    }
-
-    fn zpx(&mut self, other: &Self) {
-        self.1 ^= other.0;
-    }
-
-    fn zpz(&mut self, other: &Self) {
         self.1 ^= other.1;
     }
 
-    fn get_x(&self) -> bool {
-        self.0
+    fn xpz(&mut self, other: &Self) {
+        self.1 ^= other.0;
     }
 
-    fn get_z(&self) -> bool {
+    fn zpx(&mut self, other: &Self) {
+        self.0 ^= other.1;
+    }
+
+    fn zpz(&mut self, other: &Self) {
+        self.0 ^= other.0;
+    }
+
+    fn get_x(&self) -> bool {
         self.1
     }
 
+    fn get_z(&self) -> bool {
+        self.0
+    }
+
     fn set_x(&mut self, x: bool) {
-        self.0 = x;
+        self.1 = x;
     }
 
     fn set_z(&mut self, z: bool) {
-        self.1 = z;
+        self.0 = z;
     }
 
     fn tableau_encoding(&self) -> u8 {
-        (self.0 as u8) << 1 | self.1 as u8
+        (self.1 as u8) << 1 | self.0 as u8
     }
 }